Create transitionTest.java
The file is a JavaFX application that generates a grid game with ten balls which drop to any of the 10 spaces in the bottom of the grid. Polylines are used to create the paths and PathTranstions are used to play the transitions. Lambda coding was used for the event handlers to minimize length.